bug fixes and resolve vertical reference math (finally)
bug with converting xyz_crs to horizontal_crs, was still using the old one
changing project settings no longer alters the current_processing_status attribute
changing project settings generates processing actions based on new vert/coord system
resolve bugs with settings dialogs not reading ini file properly, not writing new settings correctly
go back to subtracting elheight, vyperdatum sep should be subtracted from the resultant
correct docstrings

+++ HSTB/kluster/gui/backends/_qgis.py
@@ -1209,7 +1209,7 @@ class MapView(QtWidgets.QMainWindow):
             the transparency of all surfaces as a percentage
         """
 
-        print('Initializing {} with transparency of {}%'.format(layername, transparency * 100))
+        print('Initializing {} with transparency of {}%'.format(layername, int(transparency * 100)))
         self.layer_background = layername
         self.layer_transparency = transparency
         self.surface_transparency = surf_transparency
